refactor!: extract manifest normalization to separate helper
