+++ src/index.ts
@@ -1,2 +1,3 @@
 export { createBundleRenderer } from './bundle-renderer'
+export { normalizeClientManifest } from './legacy'
 export { createRenderer } from './renderer'
